ref(makefile): remove manifests/ project dir
We now have a new tool - helm - for deploying these projects to Kubernetes.
These manifests have become out of date and out of sync with the upstream
charts used to install these projects.

I propose that we remove this as something mandatory for new projects. If
a project can be installed standalone, it might make sense to have something
in a contrib/ directory like contrib/kubernetes/manifests/, but it shouldn't
be a mandatory thing for introducing a new component.